@@ -3,20 +3,20 @@
  */
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs";
 
 @Injectable()
 export class CountryService {
 
   constructor(private _http: HttpClient) {
   }
-  getData(): Observable<DataBind> {
+/*  getData(): Observable<DataBind> {
    return this._http.get<DataBind[]>('/assets/country.json');
   }
 
   getDepartmentDdata() {
     return this._http.get('/assets/departmentdata.json');
-  }
+  }*/
 }
 
 export interface DataBind {
